flearn/requirements.txt
numpy
scipy
tensorflow-gpu==1.15.2
Pillow
matplotlib
tqdm
pandas